prototyping templates
