+++ scripts/dropins.js
@@ -10,75 +10,17 @@ import {
 import { initializers, Initializer } from '@dropins/tools/initializer.js';
 
 // Drop-ins
-import * as authApi from '@dropins/storefront-auth/api.js';
-import * as cartApi from '@dropins/storefront-cart/api.js';
-import * as orderApi from '@dropins/storefront-order/api.js';
+import * as Auth from '@dropins/storefront-auth/api.js';
+import * as Cart from '@dropins/storefront-cart/api.js';
+// import * as Order from '@dropins/storefront-order/api.js';
 
 // Recaptcha
 import * as recaptcha from '@dropins/tools/recaptcha.js';
 
 // Libs
-import { checkIsAuthenticated, getConfigValue, getCookie } from './configs.js';
-import { getMetadata } from './aem.js';
-import {
-  CUSTOMER_ORDER_DETAILS_PATH,
-  CUSTOMER_ORDERS_PATH,
-  ORDER_DETAILS_PATH,
-  ORDER_STATUS_PATH,
-  CUSTOMER_PATH,
-} from './constants.js';
+import { getConfigValue, getCookie } from './configs.js';
 
 export const getUserTokenCookie = () => getCookie('auth_dropin_user_token');
-
-const initializeOrderApi = (orderRef) => {
-  initializers.register(orderApi.initialize, {
-    orderRef,
-  });
-};
-
-const handleUserOrdersRedirects = () => {
-  const currentUrl = new URL(window.location.href);
-  const isAccountPage = currentUrl.pathname.includes(CUSTOMER_PATH);
-  const orderRef = currentUrl.searchParams.get('orderRef');
-  const isTokenProvided = orderRef && orderRef.length > 20;
-
-  let targetPath = null;
-  if (currentUrl.pathname.includes(CUSTOMER_ORDERS_PATH)) {
-    return;
-  }
-
-  events.on('order/error', () => {
-    if (checkIsAuthenticated()) {
-      window.location.href = CUSTOMER_ORDERS_PATH;
-    } else if (isTokenProvided) {
-      window.location.href = ORDER_STATUS_PATH;
-    } else {
-      window.location.href = `${ORDER_STATUS_PATH}?orderRef=${orderRef}`;
-    }
-  });
-
-  if (checkIsAuthenticated()) {
-    if (!orderRef) {
-      targetPath = CUSTOMER_ORDERS_PATH;
-    } else if (isAccountPage) {
-      targetPath = isTokenProvided
-        ? `${ORDER_DETAILS_PATH}?orderRef=${orderRef}`
-        : null;
-    } else {
-      targetPath = isTokenProvided
-        ? null
-        : `${CUSTOMER_ORDER_DETAILS_PATH}?orderRef=${orderRef}`;
-    }
-  } else {
-    targetPath = !orderRef ? ORDER_STATUS_PATH : null;
-  }
-
-  if (targetPath) {
-    window.location.href = targetPath;
-  } else {
-    initializeOrderApi(orderRef);
-  }
-};
 
 // Update auth headers
 const setAuthHeaders = (state) => {
@@ -118,16 +60,8 @@ const initialize = new Initializer({
 export default async function initializeDropins() {
   // Register Initializers (Global)
   initializers.register(initialize, {});
-  initializers.register(authApi.initialize, {});
-  initializers.register(cartApi.initialize, {});
-
-  // Get current page template metadata
-  const templateMeta = getMetadata('template');
-  const isOrderDetailsPage = templateMeta.includes('Order-Details');
-
-  if (isOrderDetailsPage) {
-    handleUserOrdersRedirects();
-  }
+  initializers.register(Auth.initialize, {});
+  initializers.register(Cart.initialize, {});
 
   const mount = async () => {
     // Event Bus Logger
